resume-screening-system/requirements.txt
flask
spacy
sentence-transformers
PyPDF2
pdfplumber
python-docx
detoxify
pandas
numpy
plotly
langdetect